+++ protocol/contracts/mocks/mockFacets/MockSeasonFacet.sol
@@ -132,7 +132,7 @@ contract MockSeasonFacet is SeasonFacet {
         s.season.abovePeg = num;
     }
 
-    function setLastDSoilE(uint256 number) public {
+    function setLastDSoilE(uint128 number) public {
         s.w.lastDSoil = number;
     }
 
@@ -221,7 +221,6 @@ contract MockSeasonFacet is SeasonFacet {
     function stepWeatherWithParams(
         uint256 pods,
         uint256 lastDSoil,
-        //uint256 startSoil,
         uint128 beanSown,
         uint128 endSoil,
         int256 deltaB,
@@ -231,7 +230,7 @@ contract MockSeasonFacet is SeasonFacet {
         s.season.raining = raining;
         s.r.roots = rainRoots ? 1 : 0;
         s.f.pods = pods;
-        s.w.lastDSoil = lastDSoil;
+        s.w.lastDSoil = uint128(lastDSoil);
         // s.w.startSoil = startSoil;
         s.f.beanSown = beanSown;
         s.f.soil = endSoil;
